+ fixed (mysterious) crash on loading in OSX
Don't ask me why this fixed it.  Macs are mysterious things.

+++ megamek/src/megamek/common/MiscType.java
@@ -135,7 +135,7 @@ public class MiscType extends EquipmentType {
         return 1;
     }
     
-    public float getBV(Entity entity) {
+    public double getBV(Entity entity) {
         if (bv != BV_VARIABLE) {
             return bv;
         }
